Move 'Nodes' module into a separate folder.
