+++ src/app/app-routing.module.ts
@@ -1,10 +1,10 @@
 import { NgModule } from '@angular/core';
 import { RouterModule, Routes } from '@angular/router';
-import { NodesComponent } from './nodes/nodes.component';
 import { ControlPageRootComponent } from './contorl-page/control-page-root/control-page-root.component';
 import { ControlEditingRootComponent } from './control-editing/control-editing-root/control-editing-root.component';
 import { RunsTableTabsComponent } from './runs-table-tabs/runs-table-tabs.component';
 import { MongooseSetUpComponent } from './mongoose-set-up/mongoose-set-up.component';
+import { NodesComponent } from './set-up-steps/nodes/nodes.component';
 
 const routes: Routes = [
   { path: 'nodes', component: NodesComponent },
