+++ src/app/mongoose-set-up/mongoose-set-up.component.html
@@ -15,7 +15,7 @@
       </p>
 
     
-        <div class="progress-bar bg-success" 
+        <div class="progress-bar bg-success" role="progressbar"
         [style.width.%]="tab.isCompleted ? 100 : 0"
         aria-valuenow="0" aria-valuemin="0" aria-valuemax="100">
         <!-- NOTE: Element for tab state when the set up step has been completed. -->
@@ -31,7 +31,7 @@
 </div>
 
 
-<!-- NOTE: SetUp footer  -->
+<!-- NOTE: Set Up footer  -->
 <nav id="mongoose-setup-footer" class="navbar navbar-light bg-light">
     <div class="mx-auto order-0">
         <button id="btn-continue" class="btn btn-outline-success my-2 my-sm-0" type="submit" (click)="onNextStepClicked()">
