+++ src/app/set-up-steps/nodes/nodes.component.css
