+++ src/app/set-up-steps/nodes/nodes.component.ts
@@ -1,11 +1,9 @@
 import { Component, OnInit } from '@angular/core';
-import { IpAddressService } from '../core/services/ip-addresses/ip-address.service';
-import { map, subscribeOn } from 'rxjs/operators';
-
-import { NodeConfig } from '../core/services/ip-addresses/nodeConfig';
-import { IpAddress } from '../core/services/ip-addresses/ipAddress';
+import { NodeConfig } from '../../core/services/ip-addresses/nodeConfig';
 import { Router } from '@angular/router';
 import { HttpErrorResponse } from '@angular/common/http';
+import { IpAddressService } from 'src/app/core/services/ip-addresses/ip-address.service';
+import { IpAddress } from 'src/app/core/services/ip-addresses/ipAddress';
 
 @Component({
   selector: 'app-nodes',
@@ -24,6 +22,7 @@ export class NodesComponent implements OnInit {
   constructor(private ipAddressService: IpAddressService, private router: Router) { }
 
   ngOnInit() {
+    
     this.ipAddresses = this.ipAddressService.getIpAddresses();
   }
